# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/content/{slug}": {
      "get": {
        "operationId": "PublicContentController_detail_v1",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 80,
              "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDetail"
                }
              }
            }
          }
        },
        "summary": "One published item of any type.",
        "tags": [
          "content"
        ]
      }
    },
...
    "/api/v1/admin/content": {
      "get": {
        "operationId": "AdminContentController_list_v1",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "integer",
              "minimum": 1,
              "maximum": 10000
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 25,
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "PENDING_REVIEW",
                "PUBLISHED",
                "ARCHIVED"
              ]
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "POST",
                "ANNOUNCEMENT",
                "NEWS",
                "EVENT",
                "SERMON",
                "BAPTISM"
              ]
            }
          },
          {
            "name": "branch",
            "required": false,
            "in": "query",
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "global"
                  ]
                },
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 80,
                  "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
                }
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "mine",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminContentList"
                }
              }
            }
          }
        },
        "summary": "Content you can manage, most recently changed first (all statuses).",
        "tags": [
          "admin: content"
        ]
      },
      "post": {
        "operationId": "AdminContentController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminContentDetail"
                }
              }
            }
          }
        },
        "summary": "Create a draft.",
        "tags": [
          "admin: content"
        ]
      }
    },
...
    "/api/v1/admin/content/{id}": {
      "get": {
        "operationId": "AdminContentController_get_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminContentDetail"
                }
              }
            }
          }
        },
        "tags": [
          "admin: content"
        ]
      },
      "put": {
        "operationId": "AdminContentController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminContentDetail"
                }
              }
            }
          }
        },
        "summary": "Save the editor form (status changes use the actions below).",
        "tags": [
          "admin: content"
        ]
      },
      "delete": {
        "operationId": "AdminContentController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    }
                  },
                  "required": [
                    "ok"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "summary": "Delete (soft). Authors may delete their own drafts.",
        "tags": [
          "admin: content"
        ]
      }
    },
    "/api/v1/admin/content/{id}/submit": {
      "post": {
        "operationId": "AdminContentController_submit_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminContentDetail"
                }
              }
            }
          }
        },
        "summary": "Submit a draft for review.",
        "tags": [
          "admin: content"
        ]
      }
    },
    "/api/v1/admin/content/{id}/publish": {
      "post": {
        "operationId": "AdminContentController_publish_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminContentDetail"
                }
              }
            }
          }
        },
        "summary": "Publish now or schedule for later.",
        "tags": [
          "admin: content"
        ]
      }
    },
    "/api/v1/admin/content/{id}/unpublish": {
      "post": {
        "operationId": "AdminContentController_unpublish_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminContentDetail"
                }
              }
            }
          }
        },
        "summary": "Take published content back to draft.",
        "tags": [
          "admin: content"
        ]
      }
    },
    "/api/v1/admin/content/{id}/archive": {
      "post": {
        "operationId": "AdminContentController_archive_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminContentDetail"
                }
              }
            }
          }
        },
        "tags": [
          "admin: content"
        ]
      }
    },
    "/api/v1/admin/speakers": {
      "post": {
        "operationId": "AdminSpeakersController_createSpeaker_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeakerInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlugRef"
                }
              }
            }
          }
        },
        "tags": [
          "admin: content"
        ]
      }
    },
    "/api/v1/admin/series": {
      "post": {
        "operationId": "AdminSpeakersController_createSeries_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeriesInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlugRef"
                }
              }
            }
          }
        },
        "tags": [
          "admin: content"
        ]
      }
    },
...
  },
  "info": {
    "title": "Church Platform API",
    "description": "REST API of the Church Platform. Authentication uses an HttpOnly session cookie; unsafe methods also require the CSRF token cookie echoed in the X-CSRF-Token header.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "cp_session"
      }
    },
    "schemas": {
...
      "AdminContentList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminContentRow"
            }
          },
          "page": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "pageSize": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "total": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ],
        "additionalProperties": false
      },
      "AdminContentRow": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "type": {
            "type": "string",
            "enum": [
              "POST",
              "ANNOUNCEMENT",
              "NEWS",
              "EVENT",
              "SERMON",
              "BAPTISM"
            ]
          },
          "slug": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PENDING_REVIEW",
              "PUBLISHED",
              "ARCHIVED"
            ]
          },
          "scope": {
            "type": "string",
            "enum": [
              "GLOBAL",
              "BRANCH"
            ]
          },
          "branch": {
            "nullable": true,
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "slug": {
                "type": "string"
              },
              "name": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "slug",
              "name"
            ],
            "additionalProperties": false
          },
          "publishedAt": {
            "nullable": true,
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "createdBy": {
            "nullable": true,
            "type": "string"
          },
          "isPinned": {
            "type": "boolean"
          },
          "isFeatured": {
            "type": "boolean"
          },
          "eventStartsAt": {
            "nullable": true,
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "rights": {
            "type": "object",
            "properties": {
              "edit": {
                "type": "boolean"
              },
              "submit": {
                "type": "boolean"
              },
              "publish": {
                "type": "boolean"
              },
              "archive": {
                "type": "boolean"
              }
            },
            "required": [
              "edit",
              "submit",
              "publish",
              "archive"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "id",
          "type",
          "slug",
          "path",
          "title",
          "status",
          "scope",
          "branch",
          "publishedAt",
          "updatedAt",
          "createdBy",
          "isPinned",
          "isFeatured",
          "eventStartsAt",
          "rights"
        ],
        "additionalProperties": false
      },
...
      "ContentInput": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "POST",
              "ANNOUNCEMENT",
              "NEWS",
              "EVENT",
              "SERMON",
              "BAPTISM"
            ]
          },
          "scope": {
            "type": "string",
            "enum": [
              "GLOBAL",
              "BRANCH"
            ]
          },
          "branch": {
            "nullable": true,
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "slug": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "summary": {
            "nullable": true,
            "type": "string",
            "maxLength": 500
          },
          "body": {
            "nullable": true,
            "type": "string",
            "maxLength": 100000
          },
          "authorName": {
            "nullable": true,
            "type": "string",
            "maxLength": 120
          },
          "isPinned": {
            "default": false,
            "type": "boolean"
          },
          "pinnedUntil": {
            "nullable": true,
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "isFeatured": {
            "default": false,
            "type": "boolean"
          },
          "seoTitle": {
            "nullable": true,
            "type": "string",
            "maxLength": 200
          },
          "seoDescription": {
            "nullable": true,
            "type": "string",
            "maxLength": 300
          },
          "tags": {
            "default": [],
            "maxItems": 10,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 40
            }
          },
          "event": {
            "nullable": true,
            "type": "object",
            "properties": {
              "startsAt": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
              },
              "endsAt": {
                "default": null,
                "nullable": true,
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
              },
              "allDay": {
                "default": false,
                "type": "boolean"
              },
              "timezone": {
                "default": "Africa/Johannesburg",
                "type": "string",
                "minLength": 1,
                "maxLength": 64
              },
              "category": {
                "default": "OTHER",
                "type": "string",
                "enum": [
                  "SERVICE",
                  "CONFERENCE",
                  "BAPTISM",
                  "YOUTH",
                  "PRAYER",
                  "OUTREACH",
                  "FELLOWSHIP",
                  "OTHER"
                ]
              },
              "eventStatus": {
                "default": "SCHEDULED",
                "type": "string",
                "enum": [
                  "SCHEDULED",
                  "POSTPONED",
                  "CANCELLED"
                ]
              },
              "statusNote": {
                "nullable": true,
                "type": "string",
                "maxLength": 300
              },
              "venueName": {
                "nullable": true,
                "type": "string",
                "maxLength": 200
              },
              "venueAddress": {
                "nullable": true,
                "type": "string",
                "maxLength": 300
              },
              "mapsUrl": {
                "nullable": true,
                "type": "string",
                "maxLength": 2048,
                "format": "uri"
              },
              "onlineUrl": {
                "nullable": true,
                "type": "string",
                "maxLength": 2048,
                "format": "uri"
              },
              "registrationUrl": {
                "nullable": true,
                "type": "string",
                "maxLength": 2048,
                "format": "uri"
              }
            },
            "required": [
              "startsAt"
            ]
          },
          "sermon": {
            "nullable": true,
            "type": "object",
            "properties": {
              "preachedOn": {
                "type": "string",
                "format": "date",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
              },
              "speaker": {
                "nullable": true,
                "type": "string",
                "minLength": 1,
                "maxLength": 80,
                "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
              },
              "speakerName": {
                "nullable": true,
                "type": "string",
                "maxLength": 120
              },
              "series": {
                "nullable": true,
                "type": "string",
                "minLength": 1,
                "maxLength": 80,
                "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
              },
              "scripture": {
                "nullable": true,
                "type": "string",
                "maxLength": 300
              },
              "externalVideoUrl": {
                "nullable": true,
                "type": "string",
                "maxLength": 2048,
                "format": "uri"
              },
              "durationSeconds": {
                "nullable": true,
                "type": "integer",
                "minimum": 0,
                "maximum": 86400
              },
              "language": {
                "default": "en",
                "type": "string",
                "minLength": 2,
                "maxLength": 16
              },
              "transcript": {
                "nullable": true,
                "type": "string",
                "maxLength": 200000
              }
            },
            "required": [
              "preachedOn"
            ]
          },
          "baptism": {
            "nullable": true,
            "type": "object",
            "properties": {
              "baptismDate": {
                "nullable": true,
                "type": "string",
                "format": "date",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
              },
              "candidatesCount": {
                "nullable": true,
                "type": "integer",
                "minimum": 0,
                "maximum": 10000
              },
              "officiantName": {
                "nullable": true,
                "type": "string",
                "maxLength": 120
              },
              "location": {
                "nullable": true,
                "type": "string",
                "maxLength": 200
              }
            }
          }
        },
        "required": [
          "type",
          "scope",
          "branch",
          "title"
        ]
      },
      "PublishRequest": {
        "type": "object",
        "properties": {
          "publishAt": {
            "nullable": true,
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          }
        }
      },
      "SpeakerInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "title": {
            "nullable": true,
            "type": "string",
            "maxLength": 120
          }
        },
        "required": [
          "name"
        ]
      },
      "SlugRef": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name"
        ],
        "additionalProperties": false
      },
      "SeriesInput": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "description": {
            "nullable": true,
            "type": "string",
            "maxLength": 2000
          }
        },
        "required": [
          "title"
        ]
      },
...
    }
  }
}